pyproject.toml
[project]
name = "sherlock-purple-agent"
version = "0.1.0"
description = "An autonomous cybersecurity agent for exploit generation and vulnerability analysis"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "uvicorn>=0.38.0",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
